backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.15
pydantic==1.10.7
nltk==3.8.1
spacy==3.5.3
python-dateutil==2.8.2
pandas==2.0.2
scikit-learn==1.2.2
pdfminer.six==20221105